venvTerraSense/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
distlib==0.3.7
Django==4.2.7
djongo==1.3.6
dnspython==2.4.2
filelock==3.13.1
idna==3.4
platformdirs==3.11.0
pymongo==3.7.2
pyserial==3.5
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.2.4
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.6
